fix: timers, bot

+++ bot/src/timers/GAME_START.js
@@ -1,8 +1,11 @@
+const { ObjectID } = require('mongodb');
 const { Status } = require('../constants/Game');
 const Question = require('../constants/Question');
 
 module.exports = async (client, data) => {
-	const game = await client.collections.games.findOne({ _id: data.context.id });
+	const game = await client.collections.games.findOne(new ObjectID(data.context.id));
+	if (!game || game.status !== Status.STARTING) return;
+
 	if (game.players.length < 2) {
 		await client.collections.games.deleteOne({ _id: game._id });
 
@@ -11,13 +14,15 @@ module.exports = async (client, data) => {
 		});
 	}
 
-	await client.collections.games.update({ _id: game._id }, { status: Status.AWAITING_RESPONSES });
+	await client.collections.games.updateOne({ _id: game._id }, {
+		$set: { status: Status.AWAITING_RESPONSES },
+	});
 
 	const prompt = await client.collections.questions.aggregate([
 		{
 			$match: {
 				status: Question.Status.APPROVED,
-				categories: { $in: game.categories },
+				categories: { $in: game.categories || [] },
 			},
 		},
 		{ $sample: { size: 1 } },
